Consolidate test files for feature editor UI tests.
